migration leap
- changed from partmod to bepplugin
- renamed APOFSFIX -> RK_APOFSFIX to remove annoying compiler warnings
- moved ConditionalEffects to a separate namespace and cleaned up main class
- updated assembly refs
- added .editorconfig
