+++ Region-Kit/.editorconfig
@@ -1,0 +1,6 @@
+﻿root = true
+
+[*.cs]
+indent_size = 4
+indent_style = space
+trim_trailing_whitespace = true
