+++ Region-Kit/Region-Kit/INotifyWhenRoomIsViewed.cs
@@ -5,6 +5,9 @@ using System.Text;
 
 namespace RegionKit
 {
+    /// <summary>
+    /// NOTE: notify methods registered in <see cref="MiscPO.MiscPOStatic"/>
+    /// </summary>
     public interface INotifyWhenRoomIsViewed
     {
         void RoomViewed();
