+++ Region-Kit/Region-Kit/Properties/AssemblyInfo.cs
@@ -31,6 +31,7 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.0")]
+#warning are you pushing an update? update assembly version here VVV
+[assembly: AssemblyVersion("1.4.0.0")]
 [assembly: AssemblyFileVersion(RegionKit.RegionKitMod.modVersion + "." + RegionKit.RegionKitMod.buildVersion)]
 [assembly: AssemblyInformationalVersion(RegionKit.RegionKitMod.modVersion)]
